% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_compactness_area_graphs.R
\name{calculate_attributes}
\alias{calculate_attributes}
\title{Calculate Area, Perimeter, and Compactness for Spatial Polygons}
\usage{
calculate_attributes(data)
}
\arguments{
\item{data}{An \code{sf} object containing polygon geometries.}
}
\value{
An \code{sf} object with additional columns: \code{area}, \code{perimeter}, \code{compactness}.
}
\description{
Calculate Area, Perimeter, and Compactness for Spatial Polygons
}
