% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Loading_and_checking_Buildings.R
\name{check_spatial_consistency}
\alias{check_spatial_consistency}
\title{Check spatial consistency of building datasets}
\usage{
check_spatial_consistency(buildings)
}
\arguments{
\item{buildings}{A list with two sf objects as returned by read_buildings.}
}
\value{
A list containing CRS and bounding boxes of each dataset.
}
\description{
Check spatial consistency of building datasets
}
