% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_compactness_area_graphs.R
\name{compactness_area_workflow}
\alias{compactness_area_workflow}
\title{Example Workflow for Compactness and Area Visualization}
\description{
Demonstrates how to read spatial data, calculate metrics, and visualize.
}
\examples{
\dontrun{
buildings_structured <- sf::st_read("E:/EAGLE/Second_Semester/Scientific_Graphs/Five_classes/NewCairo_structured.gpkg")
buildings_unstructured <- sf::st_read("E:/EAGLE/Second_Semester/Scientific_Graphs/Five_classes/NewCairo_unstructured.gpkg")

buildings_structured <- calculate_attributes(buildings_structured)
buildings_unstructured <- calculate_attributes(buildings_unstructured)

structured_compactness_plot <- plot_map_and_histogram(buildings_structured, "compactness", "Structured")
structured_area_plot <- plot_map_and_histogram(buildings_structured, "area", "Structured", fill_option = "D")
unstructured_compactness_plot <- plot_map_and_histogram(buildings_unstructured, "compactness", "Unstructured")
unstructured_area_plot <- plot_map_and_histogram(buildings_unstructured, "area", "Unstructured", fill_option = "D")

final_plot <- (structured_compactness_plot / structured_area_plot /
               unstructured_compactness_plot / unstructured_area_plot)
print(final_plot)
}
}
