% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Loading_and_checking_Buildings.R
\name{read_buildings}
\alias{read_buildings}
\title{Read spatial datasets}
\usage{
read_buildings(unstructured_path, structured_path)
}
\arguments{
\item{unstructured_path}{Path to the unstructured buildings GeoPackage file.}

\item{structured_path}{Path to the structured buildings GeoPackage file.}
}
\value{
A list with two sf objects: unstructured and structured buildings.
}
\description{
Read spatial datasets
}
